10 Спринт, Доработка, поправлены Service и Storage 0.2
